@@ -1,56 +1,44 @@
 name: bibli_app
 description: "A new Flutter project."
-# The following line prevents the package from being accidentally published to
-# pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-# In Windows, build-name is used as the major, minor, and patch parts
-# of the product and file versions while build-number is used as the build suffix.
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.8.1
+  sdk: '>=3.4.3 <4.0.0'
 
-# Dependencies specify other packages that your package needs in order to work.
-# To automatically upgrade your package dependencies to the latest versions
-# consider running `flutter pub upgrade --major-versions`. Alternatively,
-# dependencies can be manually updated by changing the version numbers below to
-# the latest version available on pub.dev. To see which dependencies have newer
-# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
+  cupertino_icons: ^1.0.6
+  supabase_flutter: ^2.5.6
+  shared_preferences: ^2.2.3
+  flutter_dotenv: ^6.0.0
+  cached_network_image: ^3.3.1
+  flutter_native_splash: ^2.4.0
   badges: ^3.1.2
   percent_indicator: ^4.2.3
-  shared_preferences: ^2.2.2
-  supabase_flutter: ^2.5.6
-  path_provider: ^2.1.2
-  share_plus: ^7.2.1
+  share_plus: ^7.2.2
+  just_audio: ^0.9.36
+  audio_session: ^0.1.16
   confetti: ^0.7.0
+  shimmer: ^3.0.0
+  path_provider: ^2.1.4
   http: ^1.2.2
-  just_audio: ^0.9.42
-  audio_session: ^0.1.25
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
+  lottie: ^3.1.2
+  sentry_flutter: ^8.9.0
+  package_info_plus: ^8.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^6.0.0
-
+  flutter_lints: ^3.0.0
+  mockito: ^5.4.2
+  mocktail: ^1.0.3
+  build_runner: ^2.4.7
 
 flutter:
+  uses-material-design: true
   fonts:
     - family: Poppins
       fonts:
